split `syntax` into modules
